+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Attempt to authenticate using the built-in authenticator (`authn`) with a GET
   request now results in a `404` response, rather than logging an authenticator
   not enabled message. CNJR-5854
+- Set the default and maximal limit value for resources list API to 1000 in order 
+  to align with the documentation. CNJR-8485
 
 ### Changed
 - Do not increase secret's value version in case there is no actual change in